fix: some mistakes about: register, pagination, documentation
