+++ prisma/seeds.ts
@@ -1,4 +1,4 @@
-import { PrismaClient } from "@prisma/client";
+import { Category, PrismaClient, Role } from "@prisma/client";
 
 const prisma = new PrismaClient();
 
@@ -10,7 +10,7 @@ async function main() {
       name: "Janusz",
       email: "janusz@example.com",
       password: "Sigma admin 123",
-      role: "Admin",
+      role: Role.Admin,
       isEnabled: true,
     },
   });
@@ -27,7 +27,7 @@ async function main() {
   await prisma.expense.create({
     data: {
       title: "Bilet PKP",
-      category: "Transport",
+      category: Category.Transport,
       amount: 21.37,
       date: new Date("2025-08-13"),
       trip_id: trip.trip_id,
