+++ src/app.module.ts
@@ -7,7 +7,6 @@ import { AuthModule } from "./auth/auth.module";
 import { DatabaseService } from "./database/database.service";
 import { ExpenseModule } from "./expense/expense.module";
 import { ParticipantModule } from "./participant/participant.module";
-import { RegisterModule } from "./register/register.module";
 import { TripController } from "./trip/trip.controller";
 import { TripModule } from "./trip/trip.module";
 import { TripService } from "./trip/trip.service";
@@ -17,7 +16,6 @@ import { TripService } from "./trip/trip.service";
     ParticipantModule,
     ExpenseModule,
     TripModule,
-    RegisterModule,
     AuthModule,
     ConfigModule.forRoot({ isGlobal: true }),
   ],
