+++ src/auth/auth.controller.spec.ts
@@ -1,4 +1,5 @@
-import { Test, TestingModule } from "@nestjs/testing";
+import type { TestingModule } from "@nestjs/testing";
+import { Test } from "@nestjs/testing";
 
 import { AuthController } from "./auth.controller";
 import { AuthService } from "./auth.service";
