+++ src/auth/auth.controller.ts
@@ -4,6 +4,7 @@ import { ApiOperation, ApiResponse, ApiTags } from "@nestjs/swagger";
 import { AuthService } from "./auth.service";
 import { LoginResponseDto } from "./dto/login-response.dto";
 import { LoginDto } from "./dto/login.dto";
+import { RegisterDto } from "./dto/register.dto";
 
 @ApiTags("auth")
 @Controller("auth")
@@ -25,5 +26,20 @@ export class AuthController {
   @Post("login")
   async signIn(@Body() signInDto: LoginDto): Promise<LoginResponseDto> {
     return this.authService.signIn(signInDto.email, signInDto.password);
+  }
+
+  @Post("register")
+  @HttpCode(HttpStatus.CREATED)
+  @ApiOperation({
+    summary: "Register new user",
+    description: "Register new user",
+  })
+  @ApiResponse({
+    status: 201,
+    description: "Participant registered",
+    type: RegisterDto,
+  })
+  async register(@Body() registerDto: RegisterDto) {
+    return this.authService.signUp(registerDto);
   }
 }
