+++ src/auth/dto/register.dto.ts
@@ -1,13 +1,18 @@
 import { IsEmail, IsNotEmpty } from "class-validator";
 
+import { ApiProperty } from "@nestjs/swagger";
+
 export class RegisterDto {
   @IsNotEmpty()
+  @ApiProperty()
   name: string;
 
   @IsEmail()
   @IsNotEmpty()
+  @ApiProperty()
   email: string;
 
   @IsNotEmpty()
+  @ApiProperty()
   password: string;
 }
