+++ src/participant/dto/create-participant.dto.ts
@@ -1,16 +1,21 @@
-import { IsEmail, IsNotEmpty } from "class-validator";
+import { Allow, IsEmail, IsNotEmpty } from "class-validator";
+
+import { ApiProperty } from "@nestjs/swagger";
 
 export class CreateParticipantDto {
   @IsNotEmpty()
+  @ApiProperty()
   name: string;
 
   @IsNotEmpty()
   @IsEmail()
+  @ApiProperty()
   email: string;
 
   @IsNotEmpty()
+  @ApiProperty()
   password: string;
 
-  @IsNotEmpty()
+  @Allow()
   role: string;
 }
