--- src/participant/dto/pagination.dto.ts
@@ -1,14 +1,0 @@
-import { Type } from "class-transformer";
-import { IsOptional, IsPositive, Min } from "class-validator";
-
-export class PaginationDto {
-  @IsOptional()
-  @IsPositive()
-  @Type(() => Number)
-  limit?: number;
-
-  @IsOptional()
-  @Min(0)
-  @Type(() => Number)
-  offset: number;
-}
