+++ src/participant/participant.controller.ts
@@ -2,6 +2,7 @@ import { Role } from "@prisma/client";
 import { AuthGuard } from "src/auth/auth.guard";
 import { Roles } from "src/auth/roles/roles.decorator";
 import { RoleGuard } from "src/auth/roles/roles.guard";
+import { PaginationDto } from "src/pagination/pagination.dto";
 
 import {
   Body,
@@ -22,7 +23,6 @@ import { ApiOperation, ApiResponse, ApiTags } from "@nestjs/swagger";
 
 import { CreateParticipantResponseDto } from "./dto/create-participant-response.dto";
 import { CreateParticipantDto } from "./dto/create-participant.dto";
-import { PaginationDto } from "./dto/pagination.dto";
 import { ParticipantMetadata } from "./dto/participant-metadata.dto";
 import { ParticipantUpdateResponseDto } from "./dto/participant-update-response.dto";
 import { UpdateParticipantDto } from "./dto/update-participant.dto";
