+++ src/trip/dto/create-trip.dto.ts
@@ -1,17 +1,23 @@
 import { IsDateString, IsNotEmpty, IsOptional } from "class-validator";
 
+import { ApiProperty, ApiPropertyOptional } from "@nestjs/swagger";
+
 export class CreateTripDto {
   @IsNotEmpty()
+  @ApiProperty()
   name: string;
 
   @IsNotEmpty()
   @IsDateString()
+  @ApiProperty()
   date_start: string;
 
   @IsNotEmpty()
   @IsDateString()
+  @ApiProperty()
   date_end: string;
 
   @IsOptional()
+  @ApiPropertyOptional()
   description?: string;
 }
